container_realtime/requirements.txt
fastapi
uvicorn[standard]
websockets
transformers
accelerate
numpy<2
deep-translator
google-cloud-translate
lingua-language-detector
